Move serializers to separate file, add simple health check to test app in ECS, add logger, update footer, update allowed hosts to run inside Docker container
