+++ observation_format.csv
@@ -1,0 +1,1 @@
+satellite_name,norad_cat_id,observation_time_utc,observation_time_uncertainty_sec,apparent_magnitude,apparent_magnitude_uncertainty,observer_latitude_deg,observer_longitude_deg,observer_altitude_m,Instrument,observing_mode,observing_filter,observer_email,observer_orcid,satellite_right_ascension_deg,satellite_right_ascension_uncertainty_deg,satellite_declination_deg,satellite_declination_uncertainty_deg,range_to_satellite_km,range_to_satellite_uncertainty_km,range_rate_of_satellite_km_per_sec,range_rate_of_satellite_uncertainty_km_per_sec,comments,data_archive_link,constellation
