+++ repository/models.py
@@ -4,7 +4,6 @@ from datetime import datetime
 from django.core.exceptions import ValidationError
 from django.core.validators import URLValidator
 from django.db import models
-from rest_framework import serializers
 
 
 class Satellite(models.Model):
@@ -30,12 +29,6 @@ class Satellite(models.Model):
     def save(self, *args, **kwargs):
         self.full_clean()
         super().save(*args, **kwargs)
-
-
-class SatelliteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Satellite
-        fields = ("sat_name", "sat_number", "constellation", "date_added")
 
 
 class Location(models.Model):
@@ -73,12 +66,6 @@ class Location(models.Model):
     def save(self, *args, **kwargs):
         self.full_clean()
         super().save(*args, **kwargs)
-
-
-class LocationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Location
-        fields = ("obs_lat_deg", "obs_long_deg", "obs_alt_m", "date_added")
 
 
 class Observation(models.Model):
@@ -186,37 +173,3 @@ class Observation(models.Model):
     def save(self, *args, **kwargs):
         self.full_clean()
         super().save(*args, **kwargs)
-
-
-class ObservationSerializer(serializers.ModelSerializer):
-    satellite_id = SatelliteSerializer(read_only=True)
-    location_id = LocationSerializer(read_only=True)
-
-    class Meta:
-        model = Observation
-        fields = (
-            "id",
-            "obs_time_utc",
-            "obs_time_uncert_sec",
-            "apparent_mag",
-            "apparent_mag_uncert",
-            "instrument",
-            "obs_mode",
-            "obs_filter",
-            "obs_email",
-            "obs_orc_id",
-            "sat_ra_deg",
-            "sat_ra_uncert_deg",
-            "sat_dec_deg",
-            "sat_dec_uncert_deg",
-            "range_to_sat_km",
-            "range_to_sat_uncert_km",
-            "range_rate_sat_km_s",
-            "range_rate_sat_uncert_km_s",
-            "comments",
-            "data_archive_link",
-            "flag",
-            "satellite_id",
-            "location_id",
-            "date_added",
-        )
