+++ repository/templates/repository/about.html
@@ -1,0 +1,59 @@
+{% extends "repository/base.html" %}
+{% load static %}
+
+{% block content %}
+<div class="container-fluid p-3">
+    <div class="container-md">
+      <h2 class="text-center p-2">About</h2>
+      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+        sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+         Magna ac placerat vestibulum lectus mauris ultrices eros in.
+         Libero nunc consequat interdum varius sit amet mattis vulputate
+         enim. Metus aliquam eleifend mi in nulla. In aliquam sem fringilla
+          ut morbi tincidunt augue interdum velit. Tortor condimentum
+          lacinia quis vel eros donec ac odio tempor. Sed blandit libero
+          volutpat sed cras ornare arcu dui vivamus. Tellus pellentesque
+           eu tincidunt tortor aliquam. Justo donec enim diam vulputate.
+            Erat nam at lectus urna duis. Viverra nibh cras pulvinar mattis.
+            Diam ut venenatis tellus in metus vulputate eu scelerisque felis.
+             Et molestie ac feugiat sed lectus. Eu augue ut lectus arcu.
+             Nunc sed blandit libero volutpat sed cras ornare.
+             Viverra accumsan in nisl nisi scelerisque eu. Dictum
+             fusce ut placerat orci. Magna sit amet purus gravida quis
+              blandit turpis cursus. Elit ullamcorper
+              dignissim cras tincidunt.</p>
+
+      <p>Cursus metus aliquam eleifend mi in nulla posuere sollicitudin
+        aliquam. Fermentum odio eu feugiat pretium nibh.
+        Elementum nibh tellus molestie nunc non blandit. Nunc sed
+        augue lacus viverra vitae congue eu consequat. Gravida rutrum
+        quisque non tellus orci ac auctor augue mauris. Odio euismod
+        lacinia at quis risus sed. Nec sagittis aliquam malesuada
+        bibendum arcu. Nam libero justo laoreet sit. Lectus magna
+        fringilla urna porttitor rhoncus dolor. Nulla pellentesque
+        dignissim enim sit amet venenatis urna cursus eget. Non
+        blandit massa enim nec dui nunc mattis enim. Suspendisse faucibus
+        interdum posuere lorem ipsum. Etiam non quam lacus suspendisse.</p>
+
+      <p>Feugiat in fermentum posuere urna nec tincidunt praesent.
+        Ut sem viverra aliquet eget. Ultricies mi eget mauris pharetra
+        et ultrices. Duis ut diam quam nulla porttitor. Adipiscing
+        commodo elit at imperdiet dui. Ridiculus mus mauris vitae
+        ultricies leo integer malesuada nunc. Dui accumsan sit amet
+        nulla facilisi. Pellentesque elit ullamcorper dignissim cras.
+        Purus non enim praesent elementum. Consequat semper viverra nam
+        libero justo laoreet sit amet. Vehicula ipsum a arcu cursus vitae
+        congue. Odio facilisis mauris sit amet massa vitae tortor. Turpis
+        egestas integer eget aliquet nibh. Orci porta non pulvinar neque
+        laoreet suspendisse interdum consectetur libero. Enim ut tellus
+        elementum sagittis vitae. Eu turpis egestas pretium aenean.
+        Consectetur purus ut faucibus pulvinar elementum. Nulla aliquet
+        enim tortor at auctor urna nunc. Fusce ut placerat orci nulla
+        pellentesque dignissim enim. Lacus viverra vitae congue eu consequat
+        ac felis donec et.</p>
+
+      <h3 class="text-center p-2">Acknowledgements</h3>
+        Which logos and information to include here?
+    </div>
+</div>
+{% endblock %}
