+++ requirements.txt
@@ -16,6 +16,7 @@ django-compressor==4.4
 django-libsass==0.9
 djangorestframework==3.14.0
 filelock==3.13.1
+gunicorn==21.2.0
 identify==2.5.33
 iniconfig==2.0.0
 jmespath==1.0.1
