+++ score/settings.py
@@ -12,6 +12,7 @@ https://docs.djangoproject.com/en/5.0/ref/settings/
 
 import json
 from pathlib import Path
+from socket import gethostbyname, gethostname
 
 import boto3
 from botocore.exceptions import ClientError
@@ -32,6 +33,9 @@ def get_secret(secret_name):
         # https://docs.aws.amazon.com/secretsmanager/latest/apireference/API_GetSecretValue.html
         print(e)
         raise e
+    except Exception as e:
+        print(e)
+        return None
 
     if get_secret_value_response is None:
         raise Exception("No secret value response")
@@ -47,13 +51,13 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/5.0/howto/deployment/checklist/
 
-# SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = get_secret("score-secret-key")["secret-key"]
 
-# SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
-
+DEBUG_PROPAGATE_EXCEPTIONS = True
 ALLOWED_HOSTS = []
+ALLOWED_HOSTS.append(gethostbyname(gethostname()))
+ALLOWED_HOSTS.append("127.0.0.1")
 
 
 # Application definition
@@ -100,28 +104,37 @@ TEMPLATES = [
 
 WSGI_APPLICATION = "score.wsgi.application"
 
+# SECURE_SSL_REDIRECT = True
+# SESSION_COOKIE_SECURE = True
+# CSRF_COOKIE_SECURE = True
+SECURE_BROWSER_XSS_FILTER = True
+
 
 # Database
 # https://docs.djangoproject.com/en/5.0/ref/settings/#databases
 
-DATABASES = {
-    "default": {
-        "ENGINE": "django.db.backends.postgresql_psycopg2",
-        "NAME": get_secret("score_prod_db")["dbname"],
-        "USER": get_secret("score_prod_db")["username"],
-        "PASSWORD": get_secret("score_prod_db")["password"],
-        "HOST": get_secret("score_prod_db")["host"],
-        "PORT": get_secret("score_prod_db")["port"],
-    },
-    "secondary": {
-        "ENGINE": "django.db.backends.postgresql",
-        "NAME": "score_test",
-        "USER": "postgres",
-        "PASSWORD": "postgres",
-        "HOST": "localhost",
-        "PORT": "5432",
-    },
-}
+if DEBUG:
+    DATABASES = {
+        "default": {
+            "ENGINE": "django.db.backends.postgresql",
+            "NAME": "score_test",
+            "USER": "postgres",
+            "PASSWORD": "postgres",
+            "HOST": "localhost",
+            "PORT": "5432",
+        },
+    }
+else:
+    DATABASES = {
+        "default": {
+            "ENGINE": "django.db.backends.postgresql_psycopg2",
+            "NAME": get_secret("score_prod_db")["dbname"],
+            "USER": get_secret("score_prod_db")["username"],
+            "PASSWORD": get_secret("score_prod_db")["password"],
+            "HOST": get_secret("score_prod_db")["host"],
+            "PORT": get_secret("score_prod_db")["port"],
+        },
+    }
 
 
 # Password validation
@@ -158,7 +171,7 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/5.0/howto/static-files/
 
-STATIC_URL = "/static/"
+STATIC_URL = "https://d14txihk1czyln.cloudfront.net/static/"
 STATICFILES_DIRS = [
     BASE_DIR / "static",
 ]
@@ -175,3 +188,24 @@ COMPRESS_PRECOMPILERS = (("text/x-scss", "django_libsass.SassCompiler"),)
 # https://docs.djangoproject.com/en/5.0/ref/settings/#default-auto-field
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+LOGGING = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "handlers": {
+        "console": {
+            "class": "logging.StreamHandler",
+        },
+    },
+    "root": {
+        "handlers": ["console"],
+        "level": "INFO",
+    },
+    "loggers": {
+        "django": {
+            "handlers": ["console"],
+            "level": "INFO",
+            "propagate": False,
+        },
+    },
+}
